+++ packages/types/src/global.ts
@@ -3,6 +3,7 @@
 
 import type {CosmWasmClient} from '@cosmjs/cosmwasm-stargate';
 import {CosmWasmSafeClient} from './interfaces';
+import '@subql/types-core/dist/global';
 
 declare global {
   const apiUnsafe: CosmWasmClient; //requires --unsafe flag to be defined