+++ packages/node/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Update `@subql/node-core` with fixes and support for endBlock feature (#195)
 ### Fixed
 - Missing dependencies for testing command (#194)
 - Events getting processed multiple times (#194)